+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Add `onPressed` registering on addToWallet button pressed
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Expose AddToWallet widget to render native iOS